Adds disabled code to log sequence length statics
Adds an inline boolean flag, should_log_length_stats_and_exit, which is
False and has no effect on training.

Setting this Flag to true alters behavior in the case when the command
line tool would run with streaming=False and sequence_packing=True.

In this case, when the flag is True, the command line tool will
instead read a specified sample size of shuffled items, and write a
yaml file describing the distribution of sequence length of the
sequences before packing.